FinishWithValidationForProduction

+++ Controllers/SeedController.cs
@@ -1,3 +1,4 @@
+using ClinicApi.Data;
 using ClinicApi.Services.Seed;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -10,10 +11,12 @@ namespace ClinicApi.Controllers
     public class SeedController : ControllerBase
     {
         private readonly ISeedService _seedService;
+        private readonly TokenRoles _token;
 
-        public SeedController(ISeedService seedService)
+        public SeedController(ISeedService seedService, TokenRoles token)
         {
             _seedService = seedService;
+            _token = token;
         }
         [HttpPost]
         public async Task<ActionResult<string>> Seed()
@@ -21,10 +24,21 @@ namespace ClinicApi.Controllers
             return Ok(await _seedService.SeedAll());
         }
         [HttpPost]
+        [Route("SeedEntityDept")]
+        public async Task<ActionResult<string>> SeedEntityDept()
+        {
+            return Ok(await _seedService.SeedEntityDept());
+        }
+        [HttpPost]
+        [Authorize]
         [Route("DeleteAll")]
         public async Task<ActionResult<string>> DeleteSeed()
         {
-            return Ok(await _seedService.DeleteAll());
+            var role = _token.GetRoleToken("Role");
+            if (role == "SuperAdmin")
+                return Ok(await _seedService.DeleteAll());
+            else
+                return NotFound();
         }
     }
 }
